fix!: TypeORM migrations numeric type change

+++ src/shared/infra/typeorm/migrations/1603160334980-CreateOrdersProducts.ts
@@ -22,7 +22,7 @@ export default class CreateOrdersProducts1603160334980
           },
           {
             name: 'quantity',
-            type: 'numeric',
+            type: 'integer',
           },
           {
             name: 'created_at',
